Put 1,2 and 3D strain calculation together
