+++ FEMpy/Constitutive/StrainModels/__init__.py
@@ -1,2 +1,1 @@
-from .Strain2D import *
-from .Strain3D import *
+from .ContinuumStrains import *
